fix: Fix for cells not removed (#115)
There was a logical error in processing `ChildChange` in `Graph`. When cell was removed, `ChildChange` created to change parent of removed cell to `null`. But in code, if parent is null, code to destroy child cell state was not called, so cell stayed rendered on canvas.

The fix add a check for "falsy" parent, so correct branch of code is executed and cells are destroyed.